tests/requirements.txt
flake8
pytest
pywis-pubsub
requests
paho-mqtt<2